Last commit!
This is the last commit for this project. I wasn't able to add
all of the features I initially planned on, but I was able to add
new ones, better ones to help me flex my programming muscles.
For more information including to download the .zip file, my PDFs
of my standup document and my notes for this project, I have placed
the link in my README and CHANGELOG files.